+++ usermgtwidget.h
@@ -13,19 +13,19 @@ class UserMgtWidget : public QFrame
   Q_OBJECT
 
 public:
-  explicit UserMgtWidget(int currentUserID, QWidget* parent = 0);
+  explicit UserMgtWidget(qlonglong currentUserID, QWidget* parent = 0);
   void setDatabase(QSqlDatabase db);
   bool hasAnyUserChanged() const;
 
 private:
-  const int m_currentUserID;
+  const qlonglong m_currentUserID;
   UserMgtView* m_view;
   JDatabase* m_database;
   bool m_bHasAnyUserChanged;
 
 private slots:
-  void userSelected(int id);
-  void removeUser(int id);
+  void userSelected(qlonglong id);
+  void removeUser(qlonglong id);
   void saveUser();
 };
